Refactor into redirect.py.
